# contracts/center-api.openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Monitoring Center API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /settings/audit:
    get:
      operationId: listAudit
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
        - name: project
          in: query
          schema: {type: string}
        - name: actor
          in: query
          schema: {type: string}
      responses:
        '200':
          description: Configuration changes and request outcomes, without secrets
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /projects:
    get:
      operationId: listProjects
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
        - {$ref: '#/components/parameters/Sort'}
        - {$ref: '#/components/parameters/StatusFilter'}
      responses:
        '200':
          description: Project page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProjectPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectWrite'}
      responses:
        '201': {description: Project created}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /projects/{projectId}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    put:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProjectWrite'}
      responses:
        '200': {description: Project updated}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    delete:
      operationId: disableProject
      responses:
        '204': {description: Project disabled}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /projects/{projectId}/permanent:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    delete:
      operationId: deleteProjectPermanently
      responses:
        '204': {description: Project and all dependent data permanently deleted}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /projects/{projectId}/instances:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
    get:
      operationId: listInstances
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
        - {$ref: '#/components/parameters/Sort'}
        - {$ref: '#/components/parameters/StatusFilter'}
      responses:
        '200':
          description: Instance page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstancePage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      operationId: createInstance
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InstanceWrite'}
      responses:
        '201': {description: Instance created}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /projects/{projectId}/instances/{instanceId}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/InstanceId'}
    put:
      operationId: updateInstance
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InstanceUpdate'}
      responses:
        '200': {description: Instance updated}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
    delete:
      operationId: disableInstance
      responses:
        '204': {description: Instance disabled}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /projects/{projectId}/instances/{instanceId}/permanent:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/InstanceId'}
    delete:
      operationId: deleteInstancePermanently
      responses:
        '204': {description: Instance and all dependent data permanently deleted}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /projects/{projectId}/instances/{instanceId}/resources:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/InstanceId'}
    get:
      operationId: getResources
      parameters:
        - {$ref: '#/components/parameters/Period'}
        - {$ref: '#/components/parameters/MaxPoints'}
      responses:
        '200':
          description: Resource latest and history
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ResourceResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /projects/{projectId}/instances/{instanceId}/db-pools:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/InstanceId'}
    get:
      operationId: getDbPools
      parameters:
        - {$ref: '#/components/parameters/Period'}
        - {$ref: '#/components/parameters/MaxPoints'}
      responses:
        '200':
          description: Pool latest and history
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DbPoolResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: JSESSIONID}
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    Sort:
      name: sort
      in: query
      description: Allowlisted field and direction, for example project_id,asc
      schema: {type: string}
    StatusFilter:
      name: status
      in: query
      schema: {$ref: '#/components/schemas/Status'}
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
    InstanceId:
      name: instanceId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
...
    Period:
      name: period
      in: query
      schema: {type: string, enum: [24h, 7d, 30d, 365d], default: 24h}
    MaxPoints:
      name: max_points
      in: query
      schema: {type: integer, minimum: 1, maximum: 1000, default: 1000}
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: CSRF token missing or access denied
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
...
    Identifier:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{1,63}$'
    Status:
      type: string
      enum: [UP, WARN, DOWN, UNKNOWN]
    PageMetadata:
      type: object
      required: [page, size, total]
      properties:
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 200}
        total: {type: integer, format: int64, minimum: 0}
...
    AuditEntry:
      type: object
      required: [audit_id, occurred_at, actor, action, entity_type, target_id, outcome]
      properties:
        audit_id: {type: integer, format: int64}
        occurred_at: {type: string, format: date-time}
        actor: {type: string}
        action: {type: string}
        entity_type: {type: string}
        project_id: {type: [string, 'null']}
        target_id: {type: string}
        outcome: {type: string}
        before_values: {type: [object, 'null'], additionalProperties: true}
        after_values: {type: [object, 'null'], additionalProperties: true}
    AuditPage:
      allOf:
        - {$ref: '#/components/schemas/PageMetadata'}
        - type: object
          required: [items]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/AuditEntry'}}
    Error:
      type: object
      required: [code]
      properties:
        code: {type: string}
...
    Project:
      type: object
      required: [project_id, display_name, enabled, status, total_instances, up_instances, last_seen_at]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        display_name: {type: string, maxLength: 120}
        enabled: {type: boolean}
        status: {$ref: '#/components/schemas/Status'}
        total_instances: {type: integer}
        up_instances: {type: integer}
        last_seen_at: {type: [string, 'null'], format: date-time}
    ProjectWrite:
      type: object
      required: [projectId, displayName, enabled]
      properties:
        projectId: {$ref: '#/components/schemas/Identifier'}
        displayName: {type: string, minLength: 1, maxLength: 120, pattern: '^\S(?:[\s\S]*\S)?$'}
        enabled: {type: boolean}
    ProjectPage:
      allOf:
        - {$ref: '#/components/schemas/PageMetadata'}
        - type: object
          required: [items]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/Project'}}
    Instance:
      type: object
      required: [project_id, instance_id, display_name, environment, host_name, agent_base_url,
        api_checks_enabled, poll_interval_seconds, enabled, status, system_cpu_ratio, heap_ratio,
        db_pool, internal_checks, last_seen_at]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        project_name: {type: [string, 'null']}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        instance_name: {type: [string, 'null']}
        display_name: {type: string}
        environment: {type: [string, 'null']}
        host_name: {type: [string, 'null']}
        agent_base_url: {type: string, format: uri}
        api_checks_enabled: {type: boolean}
        poll_interval_seconds: {type: integer, minimum: 15, maximum: 715827882}
        enabled: {type: boolean}
        status: {$ref: '#/components/schemas/Status'}
        system_cpu_ratio: {type: [number, 'null'], minimum: 0, maximum: 1}
        heap_ratio: {type: [number, 'null'], minimum: 0, maximum: 1}
        db_pool: {type: [string, 'null']}
        internal_checks: {type: [string, 'null']}
        last_seen_at: {type: [string, 'null'], format: date-time}
    InstanceWrite:
      type: object
      required: [instanceId, displayName, agentBaseUrl, token, apiChecksEnabled, pollIntervalSeconds, enabled]
      properties:
        instanceId: {$ref: '#/components/schemas/Identifier'}
        displayName: {type: string, minLength: 1, maxLength: 120, pattern: '^\S(?:[\s\S]*\S)?$'}
        environment: {type: [string, 'null'], maxLength: 40}
        agentBaseUrl: {type: string, format: uri, minLength: 1, maxLength: 500, description: "Normalized HTTP(S) URL without credentials, query, or fragment"}
        token: {type: string, minLength: 1, writeOnly: true, x-minUtf8Bytes: 32, description: "At least 32 bytes when UTF-8 encoded"}
        apiChecksEnabled: {type: boolean}
        pollIntervalSeconds: {type: integer, minimum: 15, maximum: 715827882}
        enabled: {type: boolean}
    InstanceUpdate:
      type: object
      required: [instanceId, displayName, agentBaseUrl, apiChecksEnabled, pollIntervalSeconds, enabled]
      properties:
        instanceId: {$ref: '#/components/schemas/Identifier'}
        displayName: {type: string, minLength: 1, maxLength: 120, pattern: '^\S(?:[\s\S]*\S)?$'}
        environment: {type: [string, 'null'], maxLength: 40}
        agentBaseUrl: {type: string, format: uri, minLength: 1, maxLength: 500, description: "Normalized HTTP(S) URL without credentials, query, or fragment"}
        token: {type: [string, 'null'], writeOnly: true, x-minUtf8BytesWhenNonBlank: 32, description: "Null or blank preserves the existing token; otherwise at least 32 bytes when UTF-8 encoded"}
        apiChecksEnabled: {type: boolean}
        pollIntervalSeconds: {type: integer, minimum: 15, maximum: 715827882}
        enabled: {type: boolean}
    InstancePage:
      allOf:
        - {$ref: '#/components/schemas/PageMetadata'}
        - type: object
          required: [items]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/Instance'}}
...
    MetricLatest:
      type: object
      additionalProperties: false
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        agent_observed_at: {type: string, format: date-time}
        status: {$ref: '#/components/schemas/Status'}
        status_reason: {type: [string, 'null']}
        central_received_at: {type: string, format: date-time}
        pid: {type: [integer, 'null'], format: int64, minimum: 1}
        jvm_start_time: {type: [string, 'null'], format: date-time}
        uptime_ms: {type: [integer, 'null'], format: int64, minimum: 0}
        system_cpu_ratio: {type: [number, 'null'], minimum: 0, maximum: 1}
        process_cpu_ratio: {type: [number, 'null'], minimum: 0, maximum: 1}
        heap_used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        heap_max_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        non_heap_used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        thread_live_count: {type: [integer, 'null'], minimum: 0}
        thread_peak_count: {type: [integer, 'null'], minimum: 0}
        gc_count: {type: [integer, 'null'], format: int64, minimum: 0}
        gc_time_ms: {type: [integer, 'null'], format: int64, minimum: 0}
        physical_memory_used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        physical_memory_total_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
    InstanceMetricSample:
      type: object
      additionalProperties: false
      required: [sampled_at, project_id, instance_id, system_cpu_ratio, process_cpu_ratio,
        heap_used_bytes, heap_max_bytes, non_heap_used_bytes, physical_memory_used_bytes,
        physical_memory_total_bytes, thread_live_count, thread_peak_count, gc_count, gc_time_ms]
      properties:
        sampled_at: {type: string, format: date-time}
        project_id: {$ref: '#/components/schemas/Identifier'}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        system_cpu_ratio: {type: [number, 'null'], minimum: 0, maximum: 1}
        process_cpu_ratio: {type: [number, 'null'], minimum: 0, maximum: 1}
        heap_used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        heap_max_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        non_heap_used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        physical_memory_used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        physical_memory_total_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        thread_live_count: {type: [integer, 'null'], minimum: 0}
        thread_peak_count: {type: [integer, 'null'], minimum: 0}
        gc_count: {type: [integer, 'null'], format: int64, minimum: 0}
        gc_time_ms: {type: [integer, 'null'], format: int64, minimum: 0}
    Disk:
      type: object
      additionalProperties: false
      required: [project_id, instance_id, path_id, path_display, used_bytes, total_bytes, status]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        path_id: {$ref: '#/components/schemas/Identifier'}
        path_display: {type: string}
        used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        total_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        status: {$ref: '#/components/schemas/Status'}
    DiskMetricSample:
      type: object
      additionalProperties: false
      required: [sampled_at, project_id, instance_id, path_id, used_bytes, total_bytes, status]
      properties:
        sampled_at: {type: string, format: date-time}
        project_id: {$ref: '#/components/schemas/Identifier'}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        path_id: {$ref: '#/components/schemas/Identifier'}
        used_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        total_bytes: {type: [integer, 'null'], format: int64, minimum: 0}
        status: {$ref: '#/components/schemas/Status'}
    ResourceResponse:
      type: object
      required: [latest, history, disks, disk_history, stale]
      properties:
        latest: {$ref: '#/components/schemas/MetricLatest'}
        history: {type: array, items: {$ref: '#/components/schemas/InstanceMetricSample'}}
        disks: {type: array, items: {$ref: '#/components/schemas/Disk'}}
        disk_history: {type: array, items: {$ref: '#/components/schemas/DiskMetricSample'}}
        stale: {type: boolean}
    DbPool:
      type: object
      additionalProperties: false
      required: [project_id, instance_id, pool_id, name, active, idle, max_size, min_idle,
        waiters, max_wait_ms, validation_latency_ms, status]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        pool_id: {$ref: '#/components/schemas/Identifier'}
        name: {type: string}
        active: {type: [integer, 'null'], minimum: 0}
        idle: {type: [integer, 'null'], minimum: 0}
        max_size: {type: [integer, 'null'], minimum: 0}
        min_idle: {type: [integer, 'null'], minimum: 0}
        waiters: {type: [integer, 'null'], minimum: 0}
        max_wait_ms: {type: [integer, 'null'], format: int64, minimum: 0}
        validation_latency_ms: {type: [integer, 'null'], format: int64, minimum: 0}
        status: {$ref: '#/components/schemas/Status'}
    DbPoolMetricSample:
      type: object
      additionalProperties: false
      required: [sampled_at, project_id, instance_id, pool_id, active, idle, max_size,
        min_idle, waiters, max_wait_ms, validation_latency_ms, status]
      properties:
        sampled_at: {type: string, format: date-time}
        project_id: {$ref: '#/components/schemas/Identifier'}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        pool_id: {$ref: '#/components/schemas/Identifier'}
        active: {type: [integer, 'null'], minimum: 0}
        idle: {type: [integer, 'null'], minimum: 0}
        max_size: {type: [integer, 'null'], minimum: 0}
        min_idle: {type: [integer, 'null'], minimum: 0}
        waiters: {type: [integer, 'null'], minimum: 0}
        max_wait_ms: {type: [integer, 'null'], format: int64, minimum: 0}
        validation_latency_ms: {type: [integer, 'null'], format: int64, minimum: 0}
        status: {$ref: '#/components/schemas/Status'}
    DbPoolResponse:
      type: object
      required: [items, history, stale]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/DbPool'}}
        history: {type: array, items: {$ref: '#/components/schemas/DbPoolMetricSample'}}
        stale: {type: boolean}
...